Move manipulations rule to file
